# backend/go/docs/swagger.yaml
basePath: /api/v2
definitions:
  blueprint:
    properties:
      id:
        type: integer
      name:
        type: string
      producedInId:
        type: integer
      producedResources:
        items:
          $ref: '#/definitions/resourceAmount'
        type: array
      productionTime:
        $ref: '#/definitions/time.Duration'
      usedResources:
        items:
          $ref: '#/definitions/resourceAmount'
        type: array
    type: object
  buildingStatus:
    enum:
      - Construction
      - Ready
      - Production
      - ResourcesNeeded
      - StorageNeeded
    type: string
    x-enum-varnames:
      - ConstructionStatus
      - ReadyStatus
      - ProductionStatus
      - ResourcesNeededStatus
      - StorageNeededStatus
  buildingType:
    properties:
      buildTime:
        $ref: '#/definitions/time.Duration'
      buildingGroup:
        type: string
      buildingSubGroup:
        type: string
      capacity:
        type: number
      cost:
        type: number
      description:
        type: string
      id:
        type: integer
      requirements:
        type: string
      title:
        type: string
      workers:
        type: integer
    type: object
  buildingWithData:
    properties:
      _id:
        type: string
      buildingType:
        $ref: '#/definitions/buildingType'
      equipment:
        items:
          $ref: '#/definitions/equipment'
        type: array
      equipmentEffect:
        items:
          $ref: '#/definitions/equipmentEffect'
        type: array
      goods:
        items:
          $ref: '#/definitions/goods'
        type: array
      hiringNeeds:
        type: integer
      level:
        type: integer
      logistics:
        $ref: '#/definitions/logistics'
      nickName:
        type: string
      onStrike:
        type: boolean
      production:
        $ref: '#/definitions/production'
      salary:
        type: number
      square:
        type: integer
      squareInUse:
        type: number
      status:
        $ref: '#/definitions/buildingStatus'
      typeId:
        type: integer
      userId:
        type: string
      workEnd:
        type: string
      workStarted:
        type: string
      workers:
        type: integer
      x:
        type: integer
      "y":
        type: integer
    type: object
...
  constructBuildingPayload:
    properties:
      square:
        type: integer
      typeId:
        type: integer
      x:
        type: integer
      "y":
        type: integer
    type: object
  equipment:
    properties:
      amount:
        type: integer
      durability:
        type: integer
      equipmentTypeId:
        type: integer
    type: object
  equipmentEffect:
    properties:
      blueprintId:
        type: integer
      effectId:
        type: integer
      value:
        type: number
      valueSecond:
        description: Second value is considered as an average for all equipment, taking
          into account the first value.
        type: number
    type: object
...
  findBuildingParams:
    properties:
      buildingTypeId:
        type: integer
      id:
        type: string
      limit:
        type: integer
      nickName:
        type: string
      order:
        type: string
      orderField:
        type: string
      page:
        type: integer
      userId:
        type: string
      x:
        type: integer
      "y":
        type: integer
    type: object
  goods:
    properties:
      price:
        type: number
      resourceTypeId:
        type: integer
      revenue:
        type: number
      sellStarted:
        type: string
      sellSum:
        type: integer
      status:
        $ref: '#/definitions/storeGoodsStatus'
    type: object
  hiringPayload:
    properties:
      buildingId:
        type: string
      hiringNeeds:
        type: integer
      salary:
        type: number
    type: object
  installEquipmentPayload:
    properties:
      amount:
        type: integer
      buildingId:
        type: string
      equipmentTypeId:
        type: integer
    type: object
  jsonResult:
    properties:
      code:
        type: integer
      data: { }
      text:
        type: string
      values: { }
    type: object
...
  logistics:
    properties:
      capacity:
        type: number
      capacityMax:
        type: number
      price:
        type: number
      revenue:
        type: number
      speed:
        type: number
    type: object
...
  production:
    properties:
      blueprintId:
        type: integer
    type: object
  resourceAmount:
    properties:
      amount:
        type: number
      resourceId:
        type: integer
    type: object
...
  startProductionPayload:
    properties:
      blueprintId:
        type: integer
      buildingId:
        type: string
      duration:
        $ref: '#/definitions/time.Duration'
    type: object
...
  storeGoodsStatus:
    enum:
      - Selling
      - DemandSatisfied
      - HighPrice
      - NotEnoughMinerals
      - SpendingLimitReached
      - CapacityReached
      - OnStrike
    type: string
    x-enum-varnames:
      - Selling
      - DemandSatisfied
      - HighPrice
      - NotEnoughMinerals
      - SpendingLimitReached
      - CapacityReached
      - OnStrike
  time.Duration:
    enum:
      - -9223372036854775808
      - 9223372036854775807
      - 1
      - 1000
      - 1000000
      - 1000000000
      - 60000000000
      - 3600000000000
    type: integer
    x-enum-varnames:
      - minDuration
      - maxDuration
      - Nanosecond
      - Microsecond
      - Millisecond
      - Second
      - Minute
      - Hour
...
host: staging.game.kube.atpstealer.com
info:
  contact: { }
  title: Game API
  version: "2.0"
paths:
  /building/blueprints:
    get:
      consumes:
        - application/json
      description: Fetches a list of blueprints. If an 'id' query parameter is provided,
        fetches the blueprint with the specified ID.
      parameters:
        - description: Blueprint ID
          in: query
          name: id
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/jsonResult'
              - properties:
                  data:
                    items:
                      $ref: '#/definitions/blueprint'
                    type: array
                type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonResult'
      summary: Get blueprints
      tags:
        - blueprints
  /building/construct:
    post:
      consumes:
        - application/json
      parameters:
        - description: Building construction payload
          in: body
          name: constructBuildingPayload
          required: true
          schema:
            $ref: '#/definitions/constructBuildingPayload'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/jsonResult'
              - properties:
                  values:
                    $ref: '#/definitions/constructBuildingPayload'
                type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jsonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonResult'
      summary: Construct a new building
      tags:
        - buildings
  /building/destroy:
    delete:
      consumes:
        - application/json
      parameters:
        - description: Building ID
          in: query
          name: _id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jsonResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jsonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonResult'
      summary: Destroy an existing building
      tags:
        - buildings
  /building/get:
    post:
      consumes:
        - application/json
      parameters:
        - description: Parameters to filter and sort buildings
          in: body
          name: findBuildingsParams
          schema:
            $ref: '#/definitions/findBuildingParams'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/jsonResult'
              - properties:
                  values:
                    items:
                      $ref: '#/definitions/buildingWithData'
                    type: array
                type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonResult'
      summary: Fetch the list of buildings
      tags:
        - buildings
  /building/hiring:
    post:
      consumes:
        - application/json
      parameters:
        - description: Details of hiring
          in: body
          name: hiringPayload
          required: true
          schema:
            $ref: '#/definitions/hiringPayload'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jsonResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jsonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonResult'
      summary: Set hiring details for a building
      tags:
        - buildings
  /building/install_equipment:
    post:
      consumes:
        - application/json
      parameters:
        - description: Equipment installation payload
          in: body
          name: installEquipmentPayload
          required: true
          schema:
            $ref: '#/definitions/installEquipmentPayload'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jsonResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jsonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonResult'
      summary: Install equipment in a building
      tags:
        - buildings
  /building/my:
    get:
      consumes:
        - application/json
      description: Optionally filter by building ID.
      parameters:
        - description: Building ID to filter by
          in: query
          name: _id
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/jsonResult'
              - properties:
                  data:
                    items:
                      $ref: '#/definitions/buildingWithData'
                    type: array
                type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jsonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonResult'
      summary: Fetch the user's buildings
      tags:
        - buildings
  /building/start_work:
    post:
      consumes:
        - application/json
      parameters:
        - description: Start production payload
          in: body
          name: startWorkPayload
          required: true
          schema:
            $ref: '#/definitions/startProductionPayload'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jsonResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jsonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonResult'
      summary: Start work in the building
      tags:
        - buildings
  /building/stop_work:
    post:
      consumes:
        - application/json
      parameters:
        - description: Production stop payload
          in: body
          name: startWorkPayload
          required: true
          schema:
            $ref: '#/definitions/startProductionPayload'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jsonResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jsonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonResult'
      summary: Stops any work in building. Later he should stop only the works available
        for stopping.
      tags:
        - buildings
  /building/types:
    get:
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/jsonResult'
              - properties:
                  data:
                    items:
                      $ref: '#/definitions/buildingType'
                    type: array
                type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonResult'
      summary: Get all building types
      tags:
        - buildings
...
swagger: "2.0"